# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Omni Model API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/voice/config": {
      "get": {
        "tags": [
          "voice"
        ],
        "summary": "Get Config",
        "operationId": "get_config_api_v1_voice_config_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceFlowConfigRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "voice"
        ],
        "summary": "Update Config",
        "operationId": "update_config_api_v1_voice_config_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceFlowConfigUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceFlowConfigRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/voice/calls": {
      "get": {
        "tags": [
          "voice"
        ],
        "summary": "Calls",
        "operationId": "calls_api_v1_voice_calls_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "call_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Call Status"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VoiceCallRead"
                  },
                  "title": "Response Calls Api V1 Voice Calls Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/voice/calls/{call_id}": {
      "get": {
        "tags": [
          "voice"
        ],
        "summary": "Get Call",
        "operationId": "get_call_api_v1_voice_calls__call_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceCallRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/voice/calls/{call_id}/review": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Review Call",
        "operationId": "review_call_api_v1_voice_calls__call_id__review_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceReviewDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceCallRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/voice/calls/{call_id}/transfer": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Transfer Call",
        "operationId": "transfer_call_api_v1_voice_calls__call_id__transfer_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceTransferRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceCallRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/voice/calls/{call_id}/recording": {
      "get": {
        "tags": [
          "voice"
        ],
        "summary": "Download Recording",
        "operationId": "download_recording_api_v1_voice_calls__call_id__recording_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "call_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/voice/simulations": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Simulate Call",
        "operationId": "simulate_call_api_v1_voice_simulations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceSimulationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceCallRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "VoiceFlowConfigRead": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "pilot_mode": {
            "type": "string",
            "enum": [
              "internal",
              "limited"
            ],
            "title": "Pilot Mode",
            "default": "internal"
          },
          "allowed_hours": {
            "additionalProperties": true,
            "type": "object",
            "title": "Allowed Hours"
          },
          "max_concurrent_calls": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Concurrent Calls",
            "default": 2
          },
          "allowed_regions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Allowed Regions"
          },
          "disclosure_text": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 20,
            "title": "Disclosure Text",
            "default": "Hello. I\u2019m an AI assistant for Omni Services. This call may be recorded. Do you consent to continue?"
          },
          "require_ai_consent": {
            "type": "boolean",
            "title": "Require Ai Consent",
            "default": true
          },
          "require_recording_consent": {
            "type": "boolean",
            "title": "Require Recording Consent",
            "default": true
          },
          "retention_days": {
            "type": "integer",
            "maximum": 3650.0,
            "minimum": 1.0,
            "title": "Retention Days",
            "default": 30
          },
          "transfer_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Transfer Number"
          },
          "emergency_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Emergency Keywords"
          },
          "allowed_tools": {
            "items": {
              "type": "string",
              "enum": [
                "find_customer",
                "create_callback_job"
              ]
            },
            "type": "array",
            "title": "Allowed Tools"
          },
          "prompt_version": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Prompt Version",
            "default": "after-hours-v1"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "flow_key": {
            "type": "string",
            "title": "Flow Key"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "flow_key",
          "created_at",
          "updated_at"
        ],
        "title": "VoiceFlowConfigRead"
      },
      "VoiceFlowConfigUpdate": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "pilot_mode": {
            "type": "string",
            "enum": [
              "internal",
              "limited"
            ],
            "title": "Pilot Mode",
            "default": "internal"
          },
          "allowed_hours": {
            "additionalProperties": true,
            "type": "object",
            "title": "Allowed Hours"
          },
          "max_concurrent_calls": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Concurrent Calls",
            "default": 2
          },
          "allowed_regions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Allowed Regions"
          },
          "disclosure_text": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 20,
            "title": "Disclosure Text",
            "default": "Hello. I\u2019m an AI assistant for Omni Services. This call may be recorded. Do you consent to continue?"
          },
          "require_ai_consent": {
            "type": "boolean",
            "title": "Require Ai Consent",
            "default": true
          },
          "require_recording_consent": {
            "type": "boolean",
            "title": "Require Recording Consent",
            "default": true
          },
          "retention_days": {
            "type": "integer",
            "maximum": 3650.0,
            "minimum": 1.0,
            "title": "Retention Days",
            "default": 30
          },
          "transfer_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Transfer Number"
          },
          "emergency_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Emergency Keywords"
          },
          "allowed_tools": {
            "items": {
              "type": "string",
              "enum": [
                "find_customer",
                "create_callback_job"
              ]
            },
            "type": "array",
            "title": "Allowed Tools"
          },
          "prompt_version": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Prompt Version",
            "default": "after-hours-v1"
          }
        },
        "type": "object",
        "title": "VoiceFlowConfigUpdate"
      },
...
      "VoiceTransferRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "VoiceTransferRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}